First commit of Quote Generation
This Quote Generator fetch an array of objects containing quotes and authors (1643 of them, due to April 9, 2021) from a remote API (https://type.fit/api/quotes) and randomly present one of them nicely to the user.

Once loaded, the user can generate another quote, again and again. Since all quotes are already set in a variable in the browser (as an Array of objects, as mentioned) there is no need to make a new GET request to the remote API in order to pull a new quote.

Users can also tweet their favorite quotes on Twitter.

This generator was built for educational purposes.

  Study Subject: JavaScript (including ES6 +), Fetch, Async / Await.