Handle with no response from remote API
In the previous commit a local file with about 10% of the quotes from the remote API was added. In the case of no response from the remote API, or any othet error might happen, the user will still be able to generate quotes from the local file and won't feel the difference.

In that case only the browser console will reveal the error.